Split all Error Result test classes
