+++ tests/Ardalis.Result.FluentAssertions.UnitTests/Constants.cs
@@ -1,0 +1,6 @@
+﻿namespace Ardalis.Result.FluentAssertions.UnitTests;
+
+public static class Constants
+{
+    public const string ErrorMessage = "Error message"; 
+}
