+++ tests/Ardalis.Result.FluentAssertions.UnitTests/InvalidResult.cs
@@ -1,0 +1,31 @@
+﻿using Xunit;
+
+namespace Ardalis.Result.FluentAssertions.UnitTests;
+
+public class InvalidResult
+{
+    [Fact]
+    public void InvalidResult_ShouldBeFailure()
+    {
+        Result.Invalid().ShouldBeFailure();
+    }
+
+    [Fact]
+    public void InvalidResultWithValidationError_ShouldBeFailure()
+    {
+        Result.Invalid(new ValidationError("IDENTIFIER", "ERROR_MESSAGE")).ShouldBeFailure();
+    }
+
+    [Fact]
+    public void InvalidResult_ShouldBeInvalid()
+    {
+        Result.Invalid().ShouldBeInvalid();
+    }
+
+    [Fact]
+    public void InvalidResultWithValidationError_ShouldBeInvalidWithValidationError()
+    {
+        var validationError = new ValidationError("IDENTIFIER", "ERROR_MESSAGE");
+        Result.Invalid(validationError).ShouldBeInvalid(validationError);
+    }
+}
